refactor(PriceTest,QuantityTest): 패키지 경로 이동
